Update LeadService.cls
As part of SIW-602, We corrected Duplicate checks  for blank and null email address fields to synchronize the duplicate checks in various parts of code.   The scenerio that was covered handled both of the following.

1) Last Name and Email

2) DOB and Email

Scenario not covered was First Name and Email (when email blank):  that is now corrected