SIW-637 Wish Design Forms not rendering
Within the flow,I disabled the Available for Input/Output on "Availability Outside Flow" within the Flow variable and the the issue was resolved.
wishVisionFormNav.cmp uses the same variables are using to access the values

I am assuming that these attributes are overriding the Flow variables.  I performed some impact analysis testing and it appears this change doesn't have impact on submitting a form via the portal or Salesforce