SIW-637 Wish Form not rendering
Within the flow,I disabled the Available for Input/Output on "Availability Outside Flow" within the Flow variable and the the issue was resolved.
Note: wishVisionFormNav.cmp uses the same variables to access the values

I am assuming that these attributes are overriding the Flow variables. I performed impact analysis testing and it appears this change doesn't have impact on submitting a form via the portal or Salesforces.

<!-- Get wish granter names variable from the flow -->
<aura:attribute name="wishGranter1" type="String"/>
<aura:attribute name="wishGranter2" type="String"/>